basic docker implementation. (Thanks Cursor) (#158)
### Description
Implemented dockerization of repository for docker use

### Reviewers
Tag reviewers.

- Required:
  
- Optional:

---
### Changelog
- Add a bulleted list of major changes

### Reviewer Guide
This is the most important part!
- No one is going to read every line of every PR, so you need to tell
the reviewers what they are looking for.
- Point out lines you want feedback on or feel unsure about.
- Highlight major changes that other members need to know about.

### Testing
#### Automatic
- Describe test cases that are covered by unit tests
#### Manual
- Describe any manual testing (launch files, visualizations, etc.)

### Documentation
- Link any relevant documentation

### Checklist
- [ ] Confirmed all tests pass on a clean build
- [ ] Added reviewers in Github
- [ ] Posted PR Summary to Discord PR's Channel
- [ ] Ran uncrustify on any modified C++ files
- [ ] Ran Colcon Lint for any modified CMakeLists.txt or Package.xml

---------

+++ scripts/launch_sim.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
-pkill -f gz
-source  $VEXU_HOME/install/setup.bash
+if [ -z "${VEXU_IN_DOCKER}" ]; then
+	pkill -f gz
+fi
+source "$VEXU_HOME/install/setup.bash"
 ros2 launch ghost_sim_examples start_sim.launch.py
